issues v0.0.1 Issues.TableView

Functions

format_for(column_widths)

Return a format string that hard codes the widths of a set of columns. We put " | " between each column.

Example

iex> widths = [5,6,99] iex> Issues.TableView.format_for(widths) “~-5s | ~-6s | ~-99s~n”

print_table_for_columns(rows, headers)

Takse a list of row data, where eacg row is a Map, and a list of headers, Pronts a table to STDOUT of the data from eacg row identified by each header. That is, eacg header identifies a column, and those columns are extracted and printed from the rows.0

We calculate the width of each column to fit the longest element in that column.

printable(str)

Return a binary (string) version of our parameter.

Examples

iex> Issues.TableView.printable(“a”) “a” iex> Issues.TableView.printable(99) “99”

puts_in_columns(data_by_columns, format)

Given a list containing rows of data, a list containing the header selectors, and a format string, write the extracted data under control of the format string.

puts_one_line_in_columns(fields, format)

separator(column_widths)

Generate the line that goes below the column headings. It is a string of hyphens, with + signs where the vertical bar between the columns goes.

Example

iex> widths = [5,6,9] iex> Issues.TableView.separator(widths) “———+————+—————“

split_into_columns(rows, headers)

Given a list of rows, where each row contains a keyed list of columns, return a list containing lists of the data in each column. The headers parameter contains the list of columns to extract

Example

iex> list = [Enum.into([{"a", "1"},{"b", "2"},{"c", "3"}], %{}), ...> Enum.into([{"a", "4"},{"b", "5"},{"c", "6"}], %{})] iex> Issues.TableView.split_into_columns(list, [ "a", "b", "c" ]) [ ["1", "4"], ["2", "5"], ["3", "6"] ]

widths_of(columns)

Given a list containing sublists, where each sublist contains the data for a column, return a list containing the maximum width of each column

Example

iex> data = [ [ "cat", "wombat", "elk"], ["mongoose", "ant", "gnu"]] iex> Issues.TableView.widths_of(data) [ 6, 8 ]